2d API_15
Camera class에 alphablend를 사용한
fadein/out 효과 구현

벽에 눈물 충돌체, 아이작과의 충돌체로 각 obj가 벽 뒤로  못넘어가게 함

눈물 터지는 효과 및 눈물 속도 느려지면 중력효과 받게함

+++ 20220302/Collider.cpp
@@ -58,10 +58,10 @@ void CCollider::render(HDC _dc)
 	Vec2 vRenderPos = CCamera::GetInst()->GetRenderPos(m_vFinalPos);
 
 	Rectangle(_dc,
-		vRenderPos.x - m_vScale.x / 2.f,
-		vRenderPos.y - m_vScale.y / 2.f,
-		vRenderPos.x + m_vScale.x / 2.f,
-		vRenderPos.y + m_vScale.y / 2.f);
+		static_cast<int>(vRenderPos.x - m_vScale.x / 2.f),
+		static_cast<int>(vRenderPos.y - m_vScale.y / 2.f),
+		static_cast<int>(vRenderPos.x + m_vScale.x / 2.f),
+		static_cast<int>(vRenderPos.y + m_vScale.y / 2.f));
 
 
 	// 함수 종료시 CSelectGDI의 소멸자 호출
